# swagger.json
{"openapi":"3.0.0","paths":{"/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/v1/users/register":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUp"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"],"security":[{"token":[]}]}},"/v1/users/login":{"post":{"operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}},"tags":["users"],"security":[{"token":[]}]}},"/v1/users":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["users"],"security":[{"token":[]}]}},"/v1/users/me":{"get":{"operationId":"UserController_aboutMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"],"security":[{"token":[]}]}},"/v1/users/{id}":{"put":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUp"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"],"security":[{"token":[]}]}},"/v1/users/{id}/role":{"put":{"operationId":"UserController_updateRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"],"security":[{"token":[]}]}},"/v1/books":{"post":{"operationId":"BookController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDto"}}}},"responses":{"201":{"description":""}},"tags":["books"],"security":[{"token":[]}]},"get":{"operationId":"BookController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["books"],"security":[{"token":[]}]}},"/v1/books/{id}":{"get":{"operationId":"BookController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["books"],"security":[{"token":[]}]},"put":{"operationId":"BookController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookUp"}}}},"responses":{"200":{"description":""}},"tags":["books"],"security":[{"token":[]}]},"delete":{"operationId":"BookController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["books"],"security":[{"token":[]}]}}},"info":{"title":"Документация по веб-сервису для приложения библиотеки","description":"Добро пожаловать в документацию по веб-сервисам нашей библиотеки. Ниже вы найдете полный список различных доступных API, а также шаблоны, используемые в нашем приложении. Подробные сведения о входах и выходах каждого API также включены, чтобы облегчить вашу интеграцию.\n  \n  Если у вас возникнут вопросы или потребуется дополнительная помощь, пожалуйста, обращайтесь к нам по адресу https://atoundji.com.\n  ","version":"1.0","contact":{}},"tags":[{"name":"cast","description":""}],"servers":[],"components":{"securitySchemes":{"token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"UserUp":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"User":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"enum":["user","administrator"],"type":"string"},"status":{"enum":["waiting","active","blocked"],"type":"string"}}},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"AuthResponse":{"type":"object","properties":{"access_token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["access_token","user"]},"UserRoleDto":{"type":"object","properties":{"role":{"enum":["user","administrator"],"type":"string"}},"required":["role"]},"BookDto":{"type":"object","properties":{"title":{"type":"string"},"author":{"type":"string"},"publicationDate":{"format":"date-time","type":"string"},"genres":{"enum":["user","administrator"],"type":"string"}},"required":["title","author","publicationDate","genres"]},"BookUp":{"type":"object","properties":{"title":{"type":"string"},"author":{"type":"string"},"publicationDate":{"format":"date-time","type":"string"},"genres":{"enum":["user","administrator"],"type":"string"}}}}}}